Use Poetry for resolving dependencies
This commit adds Poetry to use it for resolving Python dependencies instead of
`pip-compile` as I ran into some issues with a few packages using that.
`pip-sync` will still be used for installing packages into the virtual
environment, however, since `poetry` isn't capable of removing any packages
that are not specified in the `requirements.txt` or `poetry.lock` files.

Since the purpose of this repo is to get started with a new project, or just
do some experimenting, I figured the ability to automatically remove manually
installed packages would come in handy.

Also added in this commit is a `.editorconfig` file
for helping keep files nicely formatted.